Remove egress support
