+++ docs/pages/index.mdx
@@ -9,7 +9,7 @@ import { Cards } from 'nextra/components'
 | Piece | What it does |
 | --- | --- |
 | `gestaltd` | The server binary. It loads config, bootstraps plugins, serves the HTTP API, serves the web UI, and mounts `/mcp` when MCP is enabled. |
-| `config.yaml` | The declarative control plane. This is where you define auth, datastore, secrets, providers, and egress policy. |
+| `config.yaml` | The declarative control plane. This is where you define auth, datastore, secrets, providers, and UI overrides. |
 | `gestalt.lock.json` and `.gestalt/` | Optional prepared state written by `gestaltd init`. Use it for deterministic startup when you depend on remote OpenAPI, GraphQL, or packaged plugins. |
 | Deploy assets | `server/Dockerfile`, `client/Dockerfile`, and a Helm chart for production deployment paths. |
 
@@ -17,7 +17,7 @@ import { Cards } from 'nextra/components'
 
 ### Configuration
 
-Your YAML file declares the feature sets that make up a Gestalt deployment: platform auth, token storage, secret resolution, providers, and egress rules.
+Your YAML file declares the feature sets that make up a Gestalt deployment: platform auth, token storage, secret resolution, providers, and optional UI overrides.
 
 ### Prepared State
