+++ server/internal/mcpupstream/upstream_test.go
@@ -10,8 +10,6 @@ import (
 
 	"github.com/valon-technologies/gestalt/server/core"
 	"github.com/valon-technologies/gestalt/server/internal/config"
-	"github.com/valon-technologies/gestalt/server/internal/egress"
-	"github.com/valon-technologies/gestalt/server/internal/egress/egresstest"
 
 	mcpclient "github.com/mark3labs/mcp-go/client"
 	mcpgo "github.com/mark3labs/mcp-go/mcp"
@@ -439,52 +437,6 @@ func TestUpstream_RequestTokenOverridesStaticAuthorizationHeader(t *testing.T) {
 
 	ctx := WithUpstreamToken(context.Background(), "secret-token")
 	result, err := u.CallTool(ctx, "run_query", map[string]any{"sql": "SELECT 1"})
-	if err != nil {
-		t.Fatalf("CallTool: %v", err)
-	}
-	if result.IsError {
-		t.Fatalf("unexpected tool error: %v", result.Content)
-	}
-}
-
-func TestUpstream_UsesSharedEgressResolver(t *testing.T) {
-	t.Parallel()
-
-	var gotPolicy egress.PolicyInput
-	ts := newAuthenticatedHTTPTestServer(t, "Bearer secret-token")
-	t.Cleanup(ts.Close)
-
-	u, err := New(context.Background(), "clickhouse", ts.URL, core.ConnectionModeUser, nil, &egress.Resolver{
-		Subjects: egress.ContextSubjectResolver{},
-		Policy: egresstest.PolicyFunc(func(_ context.Context, input egress.PolicyInput) error {
-			gotPolicy = input
-			return nil
-		}),
-	})
-	if err != nil {
-		t.Fatalf("New: %v", err)
-	}
-
-	ctx := egress.WithSubject(context.Background(), egress.Subject{Kind: egress.SubjectUser, ID: "u1"})
-	cat, err := u.CatalogForRequest(ctx, "secret-token")
-	if err != nil {
-		t.Fatalf("CatalogForRequest: %v", err)
-	}
-	if len(cat.Operations) != 2 {
-		t.Fatalf("expected 2 operations, got %d", len(cat.Operations))
-	}
-	if gotPolicy.Subject != (egress.Subject{Kind: egress.SubjectUser, ID: "u1"}) {
-		t.Fatalf("subject = %+v, want user u1", gotPolicy.Subject)
-	}
-	if gotPolicy.Target.Method == "" {
-		t.Fatal("expected target method to be populated")
-	}
-
-	callCtx := WithUpstreamToken(
-		egress.WithSubject(context.Background(), egress.Subject{Kind: egress.SubjectUser, ID: "u1"}),
-		"secret-token",
-	)
-	result, err := u.CallTool(callCtx, "run_query", map[string]any{"sql": "SELECT 1"})
 	if err != nil {
 		t.Fatalf("CallTool: %v", err)
 	}
