+++ server/internal/provider/build.go
@@ -12,7 +12,6 @@ import (
 	coreintegration "github.com/valon-technologies/gestalt/server/core/integration"
 	"github.com/valon-technologies/gestalt/server/internal/apiexec"
 	"github.com/valon-technologies/gestalt/server/internal/config"
-	"github.com/valon-technologies/gestalt/server/internal/egress"
 	"github.com/valon-technologies/gestalt/server/internal/oauth"
 )
 
@@ -21,16 +20,11 @@ type BuildOption func(*buildOptions)
 
 type buildOptions struct {
 	authOverride coreintegration.AuthHandler
-	egress       *egress.Resolver
 }
 
 // WithAuthHandler injects a pre-built auth handler, bypassing buildAuth.
 func WithAuthHandler(h coreintegration.AuthHandler) BuildOption {
 	return func(o *buildOptions) { o.authOverride = h }
-}
-
-func WithEgressResolver(r *egress.Resolver) BuildOption {
-	return func(o *buildOptions) { o.egress = r }
 }
 
 // Build constructs a provider from a spec Definition and a ConnectionDef that
@@ -73,7 +67,6 @@ func Build(def *Definition, conn config.ConnectionDef, opts ...BuildOption) (cor
 		HTTPClient:         client,
 		Headers:            def.Headers,
 		Pagination:         buildPaginationConfigs(def),
-		EgressResolver:     bo.egress,
 	}
 
 	connMode := conn.Mode
